Implement the Crossover Operator

+++ src/PuzzleSolver/PuzzleSolver/Utilities/SettingsHelper.cs
@@ -8,7 +8,7 @@ namespace PuzzleSolver.Utilities
         internal static decimal ElitismQuotient { get; set; }
         internal static double MutationProbability { get; set; }
         internal static double DiversityQuotient { get; set; }
-        internal static double CrossoverProbability { get; set; }
+        internal static decimal CrossoverProbability { get; set; }
         internal static int TournamentSize { get; set; }
         internal static int MaxIterations { get; set; }
         internal static int OriginallySetPositions { get; set; }
